requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.41.2
openai-whisper==20231117
opencv-python==4.10.0.84
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0
shap==0.45.1
pandas==2.2.2
tqdm==4.66.4
werkzeug==3.0.3
